Fix launch assertion crash, .mins import crash, and 8-bit sample waveform display
Three independent bugs from this bug-report round (items 27-29):

1. AppDelegate.makeUntitledDocument(ofType:) asserted typeName ==
   MADNativeUTI (madk), but Cocoa-Info.plist declares many document
   types (madk, madh, madfg, every tracker format...) all sharing
   NSDocumentClass PPDocument, and AppKit is free to request an
   untitled document of any of them (e.g. window/state restoration
   remembering the type of whatever was last open). The method body
   ignores typeName entirely and always builds the same blank
   PPMusicObject, so the assert was simply wrong -- it crashed on
   ordinary launch once any non-madk document (e.g. a .madh file) had
   been open in a previous session.

2. -[PPInstrumentObject init] and -initWithCoder: both wrote through
   `writebackAddr` (via -setName:/-resetInstrument) before ever setting
   it, crashing via a NULL strlcpy destination. -initWithMusic: already
   had this exact bug fixed (writebackAddr = &theInstrument set first,
   with an explanatory comment) but -init and -initWithCoder: were
   missed. -init is what PPMINsPlug's .mins importer uses
   ([PPInstrumentObject new]), reproducing the reported crash exactly;
   -initWithCoder: is used by NSKeyedUnarchiver, including this
   session's own instrument copy/paste feature, which would have hit
   the identical crash the first time it was actually used.

3. PPSampleObject.drawSample's 8-bit rendering branch
   (PlayerPROKitAdditions.swift) reinterpreted the raw 8-bit PCM byte
   buffer as UnsafePointer<UInt16>, reading 16-bit words where every
   other index/scaling calculation in that branch assumes byte
   granularity. This read half the real samples (pairing unrelated
   bytes into bogus 16-bit values) and produced the sparse,
   disconnected-spikes waveform reported for "Kick+Tamb" (3000.madh) --
   confirmed visually before/after: gapped spikes -> a continuous,
   recognizable kick-drum waveform.

Verified live for all three: launched clean after previously opening
a .madh file, imported Kick+Tamb.mins without crashing (now instrument
000 in the list), and opened Kick+Tamb's Sample Editor to confirm the
waveform renders correctly.

+++ PlayerPROKit/PlayerPROKitAdditions.swift
@@ -369,7 +369,15 @@ extension PPSampleObject {
 			})
 		} else {
 			let sampleSize = curData.data.count
-			curData.data.withUnsafeBytes({ (theSample: UnsafePointer<UInt16>) -> Void in
+			// 8-bit PCM is one byte per sample -- theSample must be typed
+			// UInt8, not UInt16. Reinterpreting the raw byte buffer as
+			// 16-bit words here read only every other sample (each pair of
+			// unrelated 8-bit samples byte-swapped together into one bogus
+			// 16-bit value), which is exactly what produced the sparse,
+			// gapped-looking waveform reported for 8-bit samples (e.g.
+			// "Kick+Tamb" in 3000.madh) -- every index below (BS/BE/x) is
+			// already computed in byte units to match.
+			curData.data.withUnsafeBytes({ (theSample: UnsafePointer<UInt8>) -> Void in
 				var BS = start + (tSS * sampleSize) / larg
 				if isStereo {
 					BS /= 2; BS *= 2;
